+++ gcsa/_services/events_service.py
@@ -27,6 +27,8 @@ class SendUpdatesMode:
 class EventsService(BaseService):
     """Event management methods of the `GoogleCalendar`"""
 
+    _EVENTS_LIST_ORDERS = ("startTime", "updated")
+
     def _list_events(
             self,
             request_method: Callable,
@@ -174,7 +176,7 @@ class EventsService(BaseService):
         if (
                 (time_min and not isinstance(time_min, (date, datetime)))
                 or (time_max and not isinstance(time_max, (date, datetime)))
-                or (order_by and (not isinstance(order_by, str) or order_by not in self._LIST_ORDERS))
+                or (order_by and (not isinstance(order_by, str) or order_by not in self._EVENTS_LIST_ORDERS))
         ):
             raise ValueError('Calendar indexing is in the following format:  time_min[:time_max[:order_by]],'
                              ' where time_min and time_max are date/datetime objects'