+++ gcsa/event.py
@@ -1,5 +1,6 @@
 from functools import total_ordering
-from typing import List, Union
+import logging
+from typing import List, Optional, Union
 
 from beautiful_date import BeautifulDate
 from tzlocal import get_localzone_name
@@ -12,6 +13,8 @@ from .conference import ConferenceSolution, ConferenceSolutionCreateRequest
 from .person import Person
 from .reminders import PopupReminder, EmailReminder, Reminder
 from .util.date_time_util import ensure_localisation
+
+log = logging.getLogger(__name__)
 
 
 class Visibility:
@@ -44,7 +47,7 @@ class Transparency:
 class Event(Resource):
     def __init__(
             self,
-            summary: str,
+            summary: Optional[str],
             start: Union[date, datetime, BeautifulDate],
             end: Union[date, datetime, BeautifulDate] = None,
             *,
@@ -159,6 +162,13 @@ class Event(Resource):
         if isinstance(self.start, datetime) and isinstance(self.end, datetime):
             self.start = ensure_localisation(self.start, timezone)
             self.end = ensure_localisation(self.end, timezone)
+
+            if self.start.microsecond != 0 or self.end.microsecond != 0:
+                log.warning(
+                    "Microseconds are used in start/end, " +
+                    "but are not supported in the Google Calendar API " +
+                    "and will be dropped on submission."
+                )
         elif isinstance(self.start, datetime) or isinstance(self.end, datetime):
             raise TypeError('Start and end must either both be date or both be datetime.')
 
@@ -186,6 +196,12 @@ class Event(Resource):
 
         self.event_id = event_id
         self.summary = summary
+        if self.summary == "":
+            log.warning(
+                f"Summary is empty in {self}. Note that if the event is loaded "
+                + "from Google Calendar, its summary will be `None`"
+            )
+
         self.description = description
         self.location = location
         self.recurrence = ensure_list(recurrence)